Refactored login/register into User controller.
