--- HeroGame/Controllers/HeroController.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-
-namespace HeroGame.Controllers
-{
-    public class HeroController : Controller
-    {
-        // GET: Hero
-        public ActionResult Index()
-        {
-            return View();
-        }
-    }
-}
