+++ HeroGame/Controllers/HomeController.cs
@@ -29,84 +29,6 @@ namespace HeroGame.Controllers
             return View();
         }
 
-        public ActionResult LoginRegister()
-        {
-            if (Session["User"] != null)
-            {
-                Session["User"] = null;
-
-                return RedirectToAction("Index");
-            }
-            else
-            {
-                UserInfoModel model = new UserInfoModel();
-
-                return View("LoginRegister", model);
-            }
-        }
-
-        [HttpPost]
-        public ActionResult LoginRegister(UserInfoModel newUser, int logCode)
-        {
-            UserInfoModel modelUser = new UserInfoModel();
-            UserInfo_HeroModel model = new UserInfo_HeroModel();
-            IList<HeroModel> modelHeroes = new List<HeroModel>();
-
-            ViewBag.ErrorMessage = null;
-            if (logCode == 0)
-            {
-                if (ModelState.IsValid)
-                {
-                    if (userDal.CheckAvailability(newUser.Email) == false)
-                    {
-                        ViewBag.ErrorMessage = "Email has already been taken";
-
-                        return View("LoginRegister");
-
-                    }
-
-                    ViewBag.ErrorMessage = null;
-                    ViewBag.LoginError = null;
-                    userDal.SaveNewUser(newUser);
-                    modelUser = userDal.SelectUserByEmail(newUser.Email);
-                    model.User = modelUser;
-                    modelHeroes = heroDal.GetAllHeroesForUser(modelUser.Id);
-                    
-                    Session["User"] = modelUser;
-
-                    return View("Game", model);
-                }
-                else
-                {
-                    return View("LoginRegister");
-                }
-            }
-            if (logCode == 1)
-            {
-                string providedPassword = newUser.Password;
-
-                if (userDal.CheckAvailability(newUser.Email) == false)
-                {
-                    UserInfoModel user = userDal.SelectUserByEmail(newUser.Email);
-
-                    if (user.Password == providedPassword)
-                    {
-                        ViewBag.LoginError = null;
-                        ViewBag.ErrorMessage = null;
-                        model.User = userDal.SelectUserByEmail(newUser.Email);
-                        modelHeroes = heroDal.GetAllHeroesForUser(model.User.Id);                        
-                        Session["User"] = model.User;
-
-                        return View("Game", model);
-                    }
-                }
-                else
-                {
-                    ViewBag.LoginError = "Login or password was incorrect.";
-                }
-            }
-            return View("LoginRegister");
-        }
 
         public ActionResult Game()
         {
